+++ core/SCsub
@@ -183,6 +183,7 @@ SConscript("os/SCsub")
 SConscript("math/SCsub")
 SConscript("crypto/SCsub")
 SConscript("io/SCsub")
+SConscript("multiplayer/SCsub")
 SConscript("debugger/SCsub")
 SConscript("input/SCsub")
 SConscript("variant/SCsub")
