+++ core/core_constants.cpp
@@ -31,6 +31,7 @@
 #include "core_constants.h"
 
 #include "core/input/input_event.h"
+#include "core/multiplayer/multiplayer.h"
 #include "core/object/class_db.h"
 #include "core/os/keyboard.h"
 #include "core/variant/variant.h"
@@ -608,6 +609,15 @@ void register_global_constants() {
 	BIND_CORE_ENUM_CONSTANT(METHOD_FLAG_STATIC);
 	BIND_CORE_ENUM_CONSTANT(METHOD_FLAG_OBJECT_CORE);
 	BIND_CORE_ENUM_CONSTANT(METHOD_FLAGS_DEFAULT);
+
+	// rpc
+	BIND_CORE_ENUM_CONSTANT_CUSTOM("RPC_MODE_DISABLED", Multiplayer::RPC_MODE_DISABLED);
+	BIND_CORE_ENUM_CONSTANT_CUSTOM("RPC_MODE_ANY", Multiplayer::RPC_MODE_ANY);
+	BIND_CORE_ENUM_CONSTANT_CUSTOM("RPC_MODE_AUTH", Multiplayer::RPC_MODE_AUTHORITY);
+
+	BIND_CORE_ENUM_CONSTANT_CUSTOM("TRANSFER_MODE_UNRELIABLE", Multiplayer::TRANSFER_MODE_UNRELIABLE);
+	BIND_CORE_ENUM_CONSTANT_CUSTOM("TRANSFER_MODE_ORDERED", Multiplayer::TRANSFER_MODE_ORDERED);
+	BIND_CORE_ENUM_CONSTANT_CUSTOM("TRANSFER_MODE_RELIABLE", Multiplayer::TRANSFER_MODE_RELIABLE);
 
 	BIND_CORE_ENUM_CONSTANT_CUSTOM("TYPE_NIL", Variant::NIL);
 	BIND_CORE_ENUM_CONSTANT_CUSTOM("TYPE_BOOL", Variant::BOOL);
