+++ core/multiplayer/multiplayer_peer.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  multiplayer_peer_gdnative.h                                          */
+/*  multiplayer_peer.h                                                   */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,52 +28,53 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef MULTIPLAYER_PEER_GDNATIVE_H
-#define MULTIPLAYER_PEER_GDNATIVE_H
+#ifndef NETWORKED_MULTIPLAYER_PEER_H
+#define NETWORKED_MULTIPLAYER_PEER_H
 
-#include "core/multiplayer/multiplayer_peer.h"
-#include "modules/gdnative/gdnative.h"
-#include "modules/gdnative/include/net/godot_net.h"
+#include "core/io/packet_peer.h"
+#include "core/multiplayer/multiplayer.h"
 
-class MultiplayerPeerGDNative : public MultiplayerPeer {
-	GDCLASS(MultiplayerPeerGDNative, MultiplayerPeer);
+class MultiplayerPeer : public PacketPeer {
+	GDCLASS(MultiplayerPeer, PacketPeer);
 
 protected:
 	static void _bind_methods();
-	const godot_net_multiplayer_peer *interface;
 
 public:
-	MultiplayerPeerGDNative();
-	~MultiplayerPeerGDNative();
+	enum {
+		TARGET_PEER_BROADCAST = 0,
+		TARGET_PEER_SERVER = 1
+	};
 
-	/* Sets the interface implementation from GDNative */
-	void set_native_multiplayer_peer(const godot_net_multiplayer_peer *p_impl);
+	enum ConnectionStatus {
+		CONNECTION_DISCONNECTED,
+		CONNECTION_CONNECTING,
+		CONNECTION_CONNECTED,
+	};
 
-	/* Specific to PacketPeer */
-	virtual Error get_packet(const uint8_t **r_buffer, int &r_buffer_size) override;
-	virtual Error put_packet(const uint8_t *p_buffer, int p_buffer_size) override;
-	virtual int get_max_packet_size() const override;
-	virtual int get_available_packet_count() const override;
+	virtual void set_transfer_channel(int p_channel) = 0;
+	virtual int get_transfer_channel() const = 0;
+	virtual void set_transfer_mode(Multiplayer::TransferMode p_mode) = 0;
+	virtual Multiplayer::TransferMode get_transfer_mode() const = 0;
+	virtual void set_target_peer(int p_peer_id) = 0;
 
-	/* Specific to MultiplayerPeer */
-	virtual void set_transfer_channel(int p_channel) override;
-	virtual int get_transfer_channel() const override;
-	virtual void set_transfer_mode(Multiplayer::TransferMode p_mode) override;
-	virtual Multiplayer::TransferMode get_transfer_mode() const override;
-	virtual void set_target_peer(int p_peer_id) override;
+	virtual int get_packet_peer() const = 0;
 
-	virtual int get_packet_peer() const override;
+	virtual bool is_server() const = 0;
 
-	virtual bool is_server() const override;
+	virtual void poll() = 0;
 
-	virtual void poll() override;
+	virtual int get_unique_id() const = 0;
 
-	virtual int get_unique_id() const override;
+	virtual void set_refuse_new_connections(bool p_enable) = 0;
+	virtual bool is_refusing_new_connections() const = 0;
 
-	virtual void set_refuse_new_connections(bool p_enable) override;
-	virtual bool is_refusing_new_connections() const override;
+	virtual ConnectionStatus get_connection_status() const = 0;
+	uint32_t generate_unique_id() const;
 
-	virtual ConnectionStatus get_connection_status() const override;
+	MultiplayerPeer() {}
 };
 
-#endif // MULTIPLAYER_PEER_GDNATIVE_H
+VARIANT_ENUM_CAST(MultiplayerPeer::ConnectionStatus)
+
+#endif // NETWORKED_MULTIPLAYER_PEER_H
