+++ modules/gdnative/nativescript/nativescript.h
@@ -71,7 +71,7 @@ struct NativeScriptDesc {
 	};
 
 	Map<StringName, Method> methods;
-	Vector<MultiplayerAPI::RPCConfig> rpc_methods;
+	Vector<Multiplayer::RPCConfig> rpc_methods;
 	OrderedHashMap<StringName, Property> properties;
 	Map<StringName, Signal> signals_; // QtCreator doesn't like the name signals
 	StringName base;
@@ -175,7 +175,7 @@ public:
 	virtual void get_script_method_list(List<MethodInfo> *p_list) const override;
 	virtual void get_script_property_list(List<PropertyInfo> *p_list) const override;
 
-	virtual const Vector<MultiplayerAPI::RPCConfig> get_rpc_methods() const override;
+	virtual const Vector<Multiplayer::RPCConfig> get_rpc_methods() const override;
 
 	String get_class_documentation() const;
 	String get_method_documentation(const StringName &p_method) const;
@@ -213,7 +213,7 @@ public:
 	String to_string(bool *r_valid);
 	virtual Ref<Script> get_script() const;
 
-	virtual const Vector<MultiplayerAPI::RPCConfig> get_rpc_methods() const;
+	virtual const Vector<Multiplayer::RPCConfig> get_rpc_methods() const;
 
 	virtual ScriptLanguage *get_language();
