+++ modules/gdnative/pluginscript/pluginscript_instance.h
@@ -71,7 +71,7 @@ public:
 
 	void set_path(const String &p_path);
 
-	virtual const Vector<MultiplayerAPI::RPCConfig> get_rpc_methods() const;
+	virtual const Vector<Multiplayer::RPCConfig> get_rpc_methods() const;
 
 	virtual void refcount_incremented();
 	virtual bool refcount_decremented();
