+++ modules/visual_script/visual_script.h
@@ -234,7 +234,7 @@ private:
 	HashMap<StringName, Function> functions;
 	HashMap<StringName, Variable> variables;
 	Map<StringName, Vector<Argument>> custom_signals;
-	Vector<MultiplayerAPI::RPCConfig> rpc_functions;
+	Vector<Multiplayer::RPCConfig> rpc_functions;
 
 	Map<Object *, VisualScriptInstance *> instances;
 
@@ -362,7 +362,7 @@ public:
 
 	virtual int get_member_line(const StringName &p_member) const override;
 
-	virtual const Vector<MultiplayerAPI::RPCConfig> get_rpc_methods() const override;
+	virtual const Vector<Multiplayer::RPCConfig> get_rpc_methods() const override;
 
 #ifdef TOOLS_ENABLED
 	virtual bool are_subnodes_edited() const;
@@ -443,7 +443,7 @@ public:
 
 	virtual ScriptLanguage *get_language();
 
-	virtual const Vector<MultiplayerAPI::RPCConfig> get_rpc_methods() const;
+	virtual const Vector<Multiplayer::RPCConfig> get_rpc_methods() const;
 
 	VisualScriptInstance();
 	~VisualScriptInstance();
