+++ modules/websocket/doc_classes/WebSocketMultiplayerPeer.xml
@@ -32,7 +32,7 @@
 	</methods>
 	<members>
 		<member name="refuse_new_connections" type="bool" setter="set_refuse_new_connections" getter="is_refusing_new_connections" override="true" default="false" />
-		<member name="transfer_mode" type="int" setter="set_transfer_mode" getter="get_transfer_mode" override="true" enum="MultiplayerPeer.TransferMode" default="2" />
+		<member name="transfer_mode" type="int" setter="set_transfer_mode" getter="get_transfer_mode" override="true" enum="TransferMode" default="2" />
 	</members>
 	<signals>
 		<signal name="peer_packet">
